pyproject.toml
[project]
name = "voltron"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aalpy>=1.5.3",
    "chromadb>=1.3.5",
    "lxml>=6.0.2",
    "openai>=2.7.1",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
]

[[tool.uv.index]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
